init.d, install and uninstall scripts for debian

+++ scripts/install.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
-id -u leaps &>/dev/null || useradd -p leaps leaps
+id -u leaps &>/dev/null || useradd -s /bin/bash leaps
+echo leaps:leaps | chpasswd
 
 cp ./leaps /usr/sbin/leaps
-cp ./config/leaps_all_fields.js /etc/leaps.js
 cp ./config/init.d/leaps /etc/init.d/leaps
+
+if [ ! -f /etc/leaps.js ]; then
+	cp ./config/leaps_all_fields.js /etc/leaps.js
+fi
 
 chmod 755 /etc/init.d/leaps
 chown root:root /etc/init.d/leaps
